requirements.txt
Flask==3.1.0
keras==2.12.0
numpy==1.23.5
Pillow==11.0.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
